acquisition-server: First version of saving driver settings
Introduces a helper class alleviating saving and loading of variables
(properties) to and from the Configuration Manager, and subsequently
to and from the .conf file. Modified several drivers to use this method
to store their settings. Additionally modified the plugins to use the same
new mechanism. Rest of the drivers will be modified after the solution
has stabilized.