Plugins: Further fixes to the gdf reader box
Added an option to translate the signal by its physical minimum instead of the average of the physical min and max that was previously used. Using this option makes the signal look more like as it is when shown in biosig's SigViewer 0.5.2. For backwards compatibility, the default is off.
Fixed the version parsing
Tweaked the skipping of header3 padding
Updated the scenarios with the changed box
Added some verbosity to debug loglevel