Demos: Changed Motor Imagery CSP evaluation scenario streaming
For some reason, having the Stream Switch after the feature
extraction leads to worse results on the data used to train
the classifier, than if the switch was before the signal
processing chain (as after this commit).

This change is a placeholder until we have time to properly
investigate and understand what leads to the difference between
the two.


Former-commit-id: fa0823fae359748a4289b380d6a4e9fe832ad0c7