Plugins: Added DLL Bridge box to allow signal processing by libs
The DLL Bridge box allows the use of binary dlls to do signal
processing. These binary dlls can be compiled separately from
OpenViBE and used via the DLL Bridge box. To this end,
the libraries need to provide certain function signatures.

The main use-case of this box is to allow third-party
signal processing plugin development in Fortran, while
keeping the interface consisting of elementary types.
A Fortran DLL example is included.

So far, the DLL Bridge is elementary, allowing in-place editing
of signal and matrix streams by the library code.


Former-commit-id: 36fd72e3ad651c0b30f151613f62dbe7e766ee8f