Plugins: Mods to GDF boxes wrt late-occurring stimulations
- GDF Writer will now pad the signal with zeroes if it receives
stimulations that occur after the end of the signal. This is used to
handle the situation with stimulations located at the end
of the signal but the corresponding signal chunk has not been received
by the writer by the time of the box uninitialize().
- GDF Writer and Reader will warn if any stimulations are timestamped
as later than the signal end time. The reader box will still
drop such stimulations, although due to the padding feature,
the GDF Writer should no longer produce such files.


Former-commit-id: e0ca8df21f8e7c45c17a611c51e7a060801d2e1d